+++ prototypes/water_temperature/WaterTemp_CheckData.R
@@ -40,7 +40,11 @@ rm(list = ls())
 config_yaml = yaml.load_file("./wt_config.yaml")
 site = config_yaml$site
 startDate = as.Date(config_yaml$from_date)-1  # choose one day before your batch start date
+if (is.null(config_yaml$to_date)) {
+  config_yaml$to_date = format(Sys.time(), "%Y-%m-%d %H:%M:%S")
+}
 endDate = as.Date(config_yaml$to_date)+1  # choose one day after your batch end date
+
 folder_filepath = "./" # here
 
 # -----------------------------------------------------